change: move item expnansion list body to separate directory
